detect_match, detect_motion: Add two regression tests that expose a race
When calling detect_match, detect_motion, the timeout has no effect if
it occurs while the result is processed by the user code during the yield.

Note that this also affects wait_for_match and wait_for_motion as they are
implemented in terms of detect_match and detect_motion.