stbt: Fix timeout issue introduced by the branch api_enhancement
When specifying a `timeout_secs` greater than 5 seconds, the actual
timeout used was 5 seconds. This was affecting detect_match,
detect_motion, wait_for_match and wait_for_motion.

This was introduced by 411712f. The gtk timeout was not cancelled as
expected when wait_for_match was executed. The problem is that it was
expected that on_running and on_underrun were executed when detect_match
/ detect_motion was executed. However, this was not the case.

The gtk timeout is now also cancelled in `detect` and the other timeout
based on stream time is not changed.