`stbt run -vv`: Add more useful logging for detect_motion
Previously, when using `stbt run -vv`, it was not really clear what was
logged for motiondetect.

It is now clear from the name of the "absdiff" images that the mask is
not applied on these images.
Additionally, the mask that was used to detect motion on a particular
frame is logged. This will help in case the race condition recently
fixed appears again.

In order to clarify, to detect motion, first the absolute difference is
computed between two consecutive images, then the mask is applied, and
if a pixel is brighter than 40, motion is detected.
Note that in the future, we may benefit from applying an additional step
that does a threshold and erode the absolute difference.