wait_for_motion, detect_motion: Display motion detected
wait_for_motion and detect_motion now display the motiondetected in red
on the part of the screen where motion was looked for.

The part of the screen that is masked is darken. The part of the screen
where motion is looked for, is either unchanged (no motion detected) or
red (motion detected).

This will allow to visualise easily what mask is used and it will allow
to check when motion is detected.

A property `display` is added in the gstreamer element stbt-motiondetect
that defaults to true.