press_until_match accepts optional `directory` argument
So press_until_match can be called from a helper function that resides
in a different directory than the image to match. For example:

    def find_player(image):
        # some complicated logic to get to the players menu
        press_until_match("RIGHT", image,
                          directory=_caller_dir())

I wonder if we could avoid this whole issue of passing the directory
from one function to the next, by having detect_match iterate over its
callers' directories, searching for the image. One of the callers would
be the stbt script itself, so we could remove the fallback (in
detect_match) to looking in cwd. Would this approach cause any
confusion, taking the template images from places that the user might
not expect? Note that the stbt debug logs always show the absolute path
to the template image being used.