stbt: Consistent error handling: Raise RuntimeError instead of exiting
In some of these cases the stack trace from the exception won't be that
useful, because it comes from a bus message handler; but at least all
of stbt.py handles errors in the same way.

Maybe in future we will wrap everything in a try block so that we can
print pretty error messages, with the stack trace only shown when
verbose output is enabled; for now, our users will have to put up with
python's default printing of exceptions.