stbt run: Fail if wait_for_motion mask image isn't found on filesystem
In the same manner that we fail if the wait_for_match template image
isn't found.

Note that I simplified the "if" test; I think it is unlikely that any
other element will issue exactly this error message, and this way
simplifies the code a bit.

See also the notes in commit message 71071b6.