stbt run: Search for images in all the caller's directories
detect_match, wait_for_match, press_until_match, detect_motion and
wait_for_motion no longer take the optional "directory" argument;
instead, they search for the given template or mask image in their
caller's directory, then in their caller's caller's directory, etc.

Commit e5133f0 added the optional "directory" argument to
press_until_match, just like wait_for_match and detect_match already
took. This was so that press_until_match could be called from a helper
function that resides in a different directory than the image to match.
For example:

   def find_player(image):
       # some complicated logic to get to the players menu
       press_until_match("RIGHT", image,
                         directory=_caller_dir())

without the `directory` argument, press_until_match would look for
`image` in the directory containing the python script that defines
`find_player`, instead of the directory of find_player's *caller*.

The current commit avoids this whole rigmarole of passing the directory
from one function to the next, by having detect_match iterate over its
callers' directories, searching for the template image (and the same for
detect_motion when searching for the mask image).

WARNING: libraries of functions like the "find_player" example above
must take care not to contain images with commonly-used names in the
same directory as the library python file. Give the image filenames a
prefix based on the library name.

Note that this commit also adds additional checks for the existence of
the file, and raises a UITestError immediately if the file doesn't
exist. This is in addition to the checks done by the GStreamer elements,
which will emit a bus message, handled by stbt.py's
`Display.on_warning`. The new exception fails immediately and provides a
more useful stack trace (`on_warning` is called later, from the
application main loop). I've kept the checks in `on_warning` in case the
GStreamer element detects further errors with the image (maybe it exists
but isn't readable, or it isn't a valid image or mask, etc).

Note that this commit introduces the UITestError exception, for bugs in
the stbt test scripts. There is an existing UITestFailure exception for
legitimate test failures, i.e. when the tests script behaves as expected
but the system under test doesn't. (MatchTimeout and MotionTimeout are
UITestFailures.)