run-tests.sh: Print to stdout the log for failed tests
I run "make check" from a Jenkins job, and this way I can see the log
for failed tests directly in the Jenkins log (via the Jenkins web
interface) instead of having to ssh to the Jenkins server and find
the logfile in /tmp/stb-tester.XXX/log.

Right now this will help me debug a failure in
test_wait_for_match_nonexistent_template that is only happening on
Jenkins, not when I run "make check" locally.