selftests: Increase timeouts for tests expecting stbt-run to fail
Sometimes it takes several seconds for the GStreamer pipeline to launch,
so the "stbt-run" command was being timed out before it had a chance to
fail.

I haven't upped the timeouts for the tests in
test-{motiondetect,templatematch}.sh, which use a plain "gst-launch"
(instead of "stbt-run"), because those are 2 seconds we have to wait for
every single test run (as opposed to the "stbt-run" tests, where we are
expecting stbt-run to fail, and the full timeout time will only be spent
if stbt-run doesn't behave as expected). Maybe gst-launch can take more
that 2 seconds to initialise the plugin registry, get the ximagesink up
and running, etc. but I imagine that if it gets that far it's probably
going to work; those tests will still catch problems where the stbt
plugin isn't on the GStreamer search path, etc.