stbt record: also takes "-v" argument
Commit eda17239 added the "-v" flag for verbose debug output; the
intention was to add the flag to both "stbt run" and "stbt record", but
it only added it to "stbt run".

This commit makes "stbt record" also accept the "-v" argument. Now the
selftest "test_record" (in tests/test-stbt-record.sh) passes.

Also updated the README/manpage to clarify that "-v -v" only applies to
"stbt run".