README: Add (autogenerated) documentation for all user-visible python functions
Each time you modify one of the docstrings in stbt.py, the next time you
run "make doc" it will update README.rst; before committing you will
notice the change, and commit it together with your changes to stbt.py.
That's the theory, anyway. :-)

pydoc is part of the standard python distribution.

Note that "press" used to take an unspecified number of arguments and
just passed them on to the particular "control" implementation; but all
implementations only take a single argument, "key". The argument to
"press" needs to be stated explicitly now, so that it appears in the
README.

Note that I had to wrap the namedtuples (MatchResult etc) in a class so
that I could attach a docstring. See
http://stackoverflow.com/questions/1606436/