Makefile: Always generate version number from the *stb-tester* repo
I have a Jenkins server set up that builds an stb-tester RPM for me. The
configuration for that Jenkins job lives in its own git repo. The
Jenkins job was untarring the stb-tester tarball (with its VERSION file)
inside a checkout of the jenkins-rpm-job repo, so when the stb-tester
Makefile tried to determine the version from the git tags, it was
looking at the history for the jenkins-rpm-job repo instead of the
history for the stb-tester repo.

This change causes the "git describe" to fail in the scenario just
mentioned, so the stb-tester Makefile picks up the version number
from the VERSION file (shipped with the tarball) as intended.