Clarify documentation: detect_{match,motion} don't process *every* frame
but they do yield a {Match,Motion}Result object for each frame that they
*do* process. We use a leaky GStreamer queue in the pipeline, so that if
the CPU can't keep up we drop frames.

Note that we have plans to implement a "process_all_frames" function so
that, at least in a particular section of the test, you can buffer all
frames to disk and process every single frame. This will be useful for
measuring performance, the smoothness of animations, etc.