README: Document stbt's exception hierarchy
Note that I'm not documenting ConfigurationError, because that is thrown
by "stbt run" before anything in the test script could catch it; but I
am documenting UITestError, because user's own helper functions might
want to raise it.

Note that in api-doc.sh I modified the comment showing the output of
"pydoc stbt.Position", to that command's *actual* output.