Allow pipeline to restart after transient video loss during a sleep
Some background: We have a mechanism in place, based on GStreamer
underrun messages, to detect when we lose video input halfway through a
"wait_for_match" and restart the video source -- see commit b6135eca for
details. When the GStreamer pipeline never starts at all we never get
underrun messages, so we also installed a GObject timeout in commit
58285fbb; this GObject timeout is cancelled once the pipeline starts
running.

Now we've discovered a new use case in a test script that reboots the
system-under-test:

    press("POWER")
    sleep(60)  # wait for it to reboot
    wait_for_match("some-ui-element.png")  # or wait_for_motion()

Because stbt's main-loop is run synchronously inside of
wait_for_match (or wait_for_motion, etc.) instead of running all the
time in a separate thread, the pipeline-restart mechanism (after the
video loss from rebooting) only kicks in when we call wait_for_match.
The restarting takes longer than 5 seconds (see on_underrun and
restart_source_bin) and so was interrupted by the GObjectTimeout
being modified by this commit.

In other words the pipeline-restarting always worked in the *middle* of
a wait_for_match call. Now it also works at the beginning of a
wait_for_match that follows a sleep (or that follows any other non-stbt
function).

In future we really want to have the main-loop running all the time
in a separate thread; but for now this will fix the immediate problem.