Added noise_threshold parameter to wait_for_motion
The motion detect function sometimes returns false positives when dealing
with noisy analogue video signals. Addition of the noise_threshold
paramater to the motion detect function allows the script writer to
specify how tolerant of noise the motion detect function should be.

Previous to this commit, the stbt-motiondetect element calculated the
absolute difference between two consecutive frames, then applied the
mask (if given), and then applied a threshold of 40 (out of 255) -- if
any pixel was above the threshold, it was considered "motion detected".

Now, the threshold level is configurable via the element's
"noiseThreshold" property. In addition, after applying the threshold we
erode the thresholded absolute difference image to eliminate "specks" of
noise.

For further details see the discussion at
https://groups.google.com/forum/?fromgroups=#!topic/stb-tester/0ZBp_pNaHY4