LircRemote: Open a new connection to the lircd socket for each key-press
With our particular lircd.conf file (which defines the format of the
infrared signals) we would get the following error, consistently, after
68 calls to `stbt.press`:

      File "~/libexec/stbt/stbt.py", line 47, in press
        return control.press(key)
      File "~/libexec/stbt/stbt.py", line 741, in press
        self.lircd.sendall("SEND_ONCE %s %s\n" % (self.control_name, key))
      File "/usr/lib64/python2.7/socket.py", line 224, in meth
        return getattr(self._sock,name)(*args)
    socket.error: [Errno 32] Broken pipe

I don't know why this happens, or whether it is a bug in lircd. But this
commit avoids the error by opening a new connection to the lircd socket
for each call to `stbt.press`.

Now there is no need for LircRemote.close or VirtualRemote.close (the
only caller was in the unit tests). CPython's deterministic destruction
will close the connection as soon as it goes out of scope when `press`
returns.

With this change I can successfully send thousands of keypresses from a
single stbt script. The performance implications of opening a new
connection each time were so small that my measurements didn't show any
difference.