stbt support for the RedRat irNetBox network-controlled infrared emitter
For details see the previous commit, and the change to README.rst in
this commit.

Note that the irNetBox device only allows one TCP connection at a time,
so `IRNetBoxRemote.press` holds the connection for as short a time as
possible. The sleep(0.5) is unfortunate, but my observations are that
sending a second signal before waiting 0.5 seconds results in the
irNetBox successfully sending the signal but the set-top box ignoring it
completely. Setting the "delay" parameter of the irNetBox
OUTPUT_IR_ASYNC message to 500ms and removing the sleep from
`IRNetBoxRemote.press` doesn't work: Multiple calls to `press` succeed
quickly one after the other, but the set-top box only understands the
very first one. The sleep of 0.5 seconds is still short enough that
multiple `press("2")` calls will be able to enter letters in a text box,
SMS-style.

Re. the exception handling in `IRNetBoxRemote._connect` see commit
5cd6ea4a.