Remove stderr output from irnetbox python module
To allow the caller to control what output is printed.

Note that `IRNetBox.__init__` still prints an error message when it
fails to connect and will retry after a short delay, to give the user
some feedback as to why it is taking so long. It is reasonable to see
this output from `stbt run` even without the `--verbose` flag.