Instructions for OpenSUSE Build Service
We use the OpenSUSE Build Service to build packages for various systems
(currently only Fedora 16 and 17 (x86 and x86_64).

Obviously publishing a new package requires knowing the password to the
stb-tester account on the OpenSUSE Build Service. I am documenting the
procedure for my own benefit, and that of any future maintainers.