irnetbox.py: Fix connection retry on Linux
errno.ECONNREFUSED is only == 61 on my OS X system; on Linux it is 111.

This was causing stbt to give up on the first attempt at connecting to
the irNetBox, instead of retrying with an exponential backoff.

Note that retrying is necessary because the irNetBox device only allows
one TCP connection at a time, so when multiple stbt tests are using the
irNetBox simultaneously clashes are inevitable.