`make uninstall` deletes libexec/stbt/*.pyc files
Before this alteration `make uninstall` failed to delete the stbt
directory under libexec because the compiled python files (.pyc) were
still present.

This change deletes any .pyc files present in libexec (as there
are likely to be more than the 2 current files in the future) so that
`rmdir $(DESTDIR)$(libexecdir)/stbt` will work as expected.