irnetbox.py: Treat unknown irNetBox model as the MK-III
This is, frankly, for my own convenience in testing. RedRat kindly sent
me a free irNetBox-III for testing, with a faulty USB microcontroller (I
don't need the USB interface to test the network & infrared aspects of
the irNetBox, so it doesn't matter for my testing). However the USB
microcontroller is also responsible for responding to the "device
version" message, and the faulty microcontroller is saying "version =
0".

This shouldn't affect any commercial irNetBox customers out there
because "version = 0" isn't a valid version for any irNetBox model.