"make check-pylint" enforces all pep8 indentation rules
With a little bit of creativity you can achieve good readability while
still meeting the pep8 rules, and I do think consistency is a good
thing. In particular I like the rule that when using a hanging indent
there shouldn't be arguments on the first line:

    # yes:
    def long_function_name(
            var_one, var_two, var_three,
            var_four):
    # no:
    def long_function_name(var_one, var_two, var_three,
            var_four):

I did have to silence E128 (continuation line under-indented for visual
indent) for irnetbox.py, because it interfered with the config file
specified inline as a string in the unit test
test_that_parse_config_understands_redrat_format.