New function `stbt.get_config` for stbt scripts to read from stbt.conf
At YouView we have been writing libraries of helper functions, like
"network.disable()" to control the network connectivity of our
system-under-test. So far we have been using environment variables to
let those libraries know, for example, the network interface of the
system-under-test, so that we can run the same stbt script and the same
library, without modification, against multiple different set-top boxes.

So we've ended up with some information in stbt.conf (e.g.
source-pipeline, which is different for each system-under-test because
the "/dev/videoN" device is different) and some information in
environment variables.

Thanks to this new `stbt.get_config` function we can store all the
information in stbt.conf, and have our custom libraries read from it.
Then if we have multiple set-top-boxes connected to a single PC, we
create a different stbt.conf for each one, and set STBT_CONFIG_FILE
appropriately in the environment of `stbt run`.

Note that `stbt.load_config` only reads the config files once, at the
beginning of `stbt run`, to ensure that the stbt script sees a
consistent view of the configuration files for the duration of its run.

Note that (unlike `stbt.wait_for_match` etc) stbt-run doesn't import
`get_config` into the top-level namespace of the stbt script, because
we want to encourage people to import dependencies into their scripts
explicitly (see commit 35bbf455).