stbt-templatematch, wait_for_match, et al. are fully configurable by user
This patch allows users full configurability of the templatematch
algorithm through all aspects of stb-tester package which imply
matching, i.e.
  * stbt-templatematch (gstreamer plugin)
  * stbt.wait_for_match(...)
  * stbt.press_until_match(...)
  * stbt.detect_match(...)

These parameters are:
gstElement param | stbt.py param
_________________|__________________
matchMethod      | match_method
matchThreshold   | match_threshold
confirmMethod    | confirm_method
erodePasses      | erode_passes
confirmThreshold | confirm_threshold*

* Previously known as noiseThreshold/noise_threshold. noiseThreshold is
  marked for deprecation and usage of it will raise a warning. At
  present both noise_threshold and confirm_threshold are valid
  individually when set in wait_for_match (*not* stbt-templatematch
  element though!), and affect the template match in the same way.
  However, using both will result in an error.

For full details on these parameters, please read the documentation,
see stbt's man page, or use gst-inspect stbt-templatematch.

Selftests have been implemented to ensure that setting each of the
parameters to stbt-templatematch are accepted by the element and have
an effect on the template matching process.

Each of the tests that execute a `wait_for_match` are done as 2
consecutive scripts executed with stbt-run. This is to prevent a known
race condition (see commit 243c27e6) where the second in a pair of
consecutive detect_matches reports a match falsely because of a second
bus message from the gstreamer element which reports the positive match
from the first detect_match that stbt.py has already acted on, and which
stbt interprets as a successful match for the second detect_match. There
is a mechanism implemented to make stbt ignore these superfluous
messages, however it only takes effect if the 2 subsequent
detect_matches are using different templates. Because these tests are
designed to demonstrate the effect of changing a single templatematch in
a test, they must use the same template. (They could use identical
copies of each template only with a different name, however this would
make it confusing to understand the tests' purpose and needlessly
increase the number of files in this package.)