stbt templatematch: New command-line tool to compare two images
This is useful for experimenting with different `wait_for_match`
parameters, and for writing unit tests for your stbt scripts.

(Whenever we write a stbt script that has to override the wait_for_match
`noise_threshold`, we capture a screenshot and write a unit test —using
`stbt templatematch`— that will fail if the image comparison starts
matching _without_ the custom `noise_threshold`. That way we'll know to
remove the custom `noise_threshold` if the templatematch algorithm is
improved in a future release of stbt. The unit test also serves as
documentation for why the custom `noise_threshold` was necessary.)