stbt-completion: Cleanup: Remove initial "#!bash" line
It isn't actually a valid interpreter directive, because it doesn't have
an absolute path. It *shouldn't* be an interpreter directive because the
script is intended to be sourced into your shell, not run as a
subprocess; but I think it could be confusing to some people because it
looks *almost* like an interpreter directive.

It did serve one purpose: Telling editors like Emacs what type of file
this is, for syntax highlighting etc. The new "-*- shell-script -*-"
has the same effect.