New stbt-templatematch confirm method: 'normed-absdiff'
This is a third option for `stbt-templatematch`'s confirmMethod and
`wait_for_match`'s confirm_method parameter in addition to the existing:
'none' and 'absdiff' methods.

'normed-absdiff' is a slight modification and extension of 'absdiff'. The
only difference is a normalise step on both the template and source before
the absolute difference is done. The ethos behind this step came when we
started finding a number of matches that were giving false positives due
to the template image having a low dynamic range, i.e. being low contrast.
By normalizing template and source we exaggerate any differences between
them, meaning it is less likely for a false match to happen.

However, this also means the test is more susceptible to noise, and
therefore the default confirm_threshold (noise_threshold) has been
increased from 0.16 to 0.28. This is preferable over doubling the number
of erode_passes because it is much easier to fine tune the
confirm_threshold than the erode_passes.

The selftest which checked the effect of changing confirm_threshold has
been updated, and a new test to check the effect of switching between
confirm methods is implemented, which is based on a false match we found.

Some of the selftests had to be changed to explicitly use the old
template matching algorithm and settings. i.e. confirm_method=1,
confirm_threshold=0.16, because they *do not* pass with the normalize method,
which is now default.