fixes for pylint error compliance
This change (1) removes the --errors-only flag from the pylint checker
so that now pylint will warn about refactors, conventions, and warnings
and not just errors, and (2) applies all of the pylint recommendations
to the affected python modules.

Some sensible global pylint exceptions as well as module-specific
exceptions are added in `extra/pylint.sh`. In acceptable situations,
inline pylint "disable" codes are added in the code.

A notable change for compliance is that the stbt.py `display` and
`control` variables are now module-private variables `_display` and
`_control`.