gsttemplatematch.c: using non-normalized `matchMethod` raises a warning
stb-tester's template match "supports" the non-normalized methods
OpenCV uses for template matching, however the resulting "heat maps" do
not make sense within the context of STBT and the `matchThreshold`
parameter. What is meant by this, is that `matchThreshold` is limited
to the range [0.0..1.0f], but the pixel values for the non-normalized
heat maps do not have a concrete upper limit, and almost definitely fall
outside the [0.0..1.0f] range.

To this end, using a non-normalized method will almost always pass through
to the second pass as the `matchThreshold` would always be exceeded, and
therefore it is *strongly* recommended (but not enforced) that the
equivalent normalized version of the `matchMethod` is used instead.