`match_method`/`confirm_method` params take literal string names not ints
Before, the 2 method parameters accepted integer values which were the
equivalent to the enumerated values in the C code. Now, they take literal
strings of the actual names of the desired method. This is much clearer to
read and interpret than just numbers.

Note: when passed as a parameter via the Python API, the methods must be
string literals (i.e. single or double quoted), however in stbt.conf files,
the quotes must be omitted. When passed as a command line argument to
stbt-templatematch, the quotes must be escaped (i.e. \"...\") as in
test-stbt-templatematch.sh so that the quotes get parsed literally to the
program.

Note that the stbt-templatematch gstreamer element has always accepted the
method names *as well* as the enumerated values, however this modification
means that with the Python API the method names are now accepted *instead
of* the enumerated values.