Fix false-pass of test_gstreamer_core_elements
If the `gst-launch` command runs successfully but the pipeline never
enters the PLAYING state, the test used to pass.

This happened on OS X if I didn't have an X11 server running. This
particular test would pass, but subsequent tests would fail. The point
of this test is to catch such environmental issues.