Ensure GStreamer tests test the source plugin, not the installed plugin
GStreamer has a "registry" (a cache of element names and the path of the
shared library implementing each element) so if the `stbt-templatematch`
element was already in the registry, `gst-launch` and `gst-inspect`
would use the element from the registry (most likely from an installed
version of stbt) instead of obeying $GST_PLUGIN_PATH, which is set by
`run-tests.sh` to look in the source directory.

Deleting the registry has no harmful effects other than the next
invocation of `gst-launch` or `gst-inspect` will take slightly longer to
start as it recreates the registry.

There's no need to delete the registry again after running the tests,
because when a user runs the installed version of `stbt`, he/she won't
have GST_PLUGIN_PATH pointing at the source directory; even if the
plugin is in the registry, it won't be used if not on GStreamer's search
path.