tests/stbt.conf: Add `sink_pipeline` removed by error in 01d25dac
The symptom was that the tests from tests/test-stbt-*.sh didn't display
on screen if the user's own ~/.config/stbt/stbt.conf had
`sink_pipeline = fakesink`. Perhaps that is actually a more desirable
behaviour, but it needs to be a conscious decision, not a side-effect
of 01d25dac.

I'm also removing `global.control`, because 01d25dac added `run.control`
and `record.control`. Some tests for `stbt.get_config` and `stbt config`
used the `global.control` key but only because it was conveniently
there; I've changed those tests to use `global.not_special`, a config
key specifically dedicated to those tests.