Remove README.rst.in (`make doc` keeps README.rst up-to-date)
Ever since 2204c898, `make doc` updates README.rst (in place) so that
the python API documentation is kept in sync with the docstrings in
`stbt.py`.

882f6b76 added additional information to this process; it's no longer
just the python API documentation, it's also the default values of the
`wait_for_match` parameters (which are now not specified in the python
code, but in `stbt.conf`).

However the *process* by which 882f6b76 updates `README.rst` is
different than the existing process for updating the python api docs in
the same file. This current commit uses the same process for updating
both sections of the documentation. The advantage is that it removes the
almost-identical `README.rst.in` added by 882f6b76, thus removing the
following inconsistency:

                         README.rst.in         README.rst

    python api docs      Contained full but    Contains full python
                         obsolete python api   api docs,
                         docs, *not* updated   updated by `make doc`.
                         by `make doc`.

    stbt.conf defaults   Contained             Contains the values,
                         placeholder.          updated by `make doc`.

It was already somewhat non-standard to commit a generated file
(`README.rst`). Having both `README.rst` and `README.rst.in` in the
repository is even more confusing.