stbt.py: Wording of `noise_threshold` deprecation warning
Removed "stbt-templatematch" from the deprecation message.

"stbt-templatematch" is the name of the GStreamer element used behind
the scenes by `stbt.wait_for_match` et al.

`noise_threshold` doesn't exist at all in the GStreamer element, so the
message is misleading. Furthermore, users of the python API don't need
to know about the GStreamer element --it's an implementation detail-- so
the name of the GStreamer element shouldn't be surfaced in a
user-visible message.

This also changes the `filterwarnings` setup to print all deprecation
warnings originating from the `stbt` module; but this is the only
warning raised by the module.