The method FramProcessor::get_elapsed_time() has been renamed for more clearance on the used metrics (seconds).
An initial image has been added when processing all frames so as to keep
track of a transition, and calculate the average speed, and perhaps in
future versions, more specific values.

Notes:

1) In order to calculate the speed of a transition or the distance, an
initial and final images must be specified. Otherwise, there wouldn't be
a start or end reference to take.

2) As an example of usage of the changes made so far:

    with stbt.process_all_frames(
                          initial_image='first.png',
                          last_image='second.png') as frame_processor:
        frame_processor.perform_action(lambda: stbt.press('CURSOR_RIGHT'))
        print 'Time:', frame_processor.get_elapsed_seconds()
        print 'Distance moved:', frame_processor.get_distance_moved()
        print 'Average speed:', frame_processor.get_average_speed()